v2.0.0

+++ chat/views.py
@@ -2,11 +2,10 @@ from rest_framework import status
 from rest_framework.views import APIView
 from rest_framework.generics import get_object_or_404
 from rest_framework.viewsets import ViewSet
-from rest_framework import status
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
-from .models import ChatRoom, RoomMessage
-from .serializers import MessageSerializer, ChatRoomSerializer
+from .models import ChatRoom, RoomMessage, RoomChatParticipant
+from .serializers import ParticipantSerializer, MessageSerializer, ChatRoomSerializer
 
 
 class ChatViewSet(ViewSet):
@@ -19,8 +18,8 @@ class ChatViewSet(ViewSet):
         return Response(serializer.data, status=status.HTTP_200_OK) 
     
     # 특정 방의 요청이 'retrive' 오면 채팅방의 채팅 보여주기
-    def retrieve(self, request, pk=None): # pk = room_id
-        room = get_object_or_404(ChatRoom, pk=pk)
+    def retrieve(self, request, room_id=None): # pk = room_id
+        room = get_object_or_404(ChatRoom, pk=room_id)
         self.check_object_permissions(request, room)
         RoomMessage.objects.filter(room_id=room.id).exclude(author_id = request.user.id).update(
             is_read=True
@@ -31,20 +30,43 @@ class ChatViewSet(ViewSet):
         return Response(serializer.data, status=status.HTTP_200_OK) 
 
 
-# 채팅방 생성(삭제)
+# 채팅방 생성(삭제), 특정 채팅방 정보 보여주기
 class ChatRoomView(APIView):
+    permission_classes = [IsAuthenticated]
+    
+    def get(self, request, room_id):
+        room = get_object_or_404(ChatRoom, id=room_id)
+        room_serializer = ChatRoomSerializer(room)
+        participants = RoomChatParticipant.objects.filter(room_id=room_id)
+        participants_serializer = ParticipantSerializer(participants, many=True)
+        
+        data = {
+            "room":room_serializer.data,
+            "participants":participants_serializer.data
+        }
+        return Response(data, status=status.HTTP_200_OK)
+        
     def post(self, request):
-        serializer = ChatRoomSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(status=status.HTTP_201_CREATED)
+        queryset = ChatRoom.objects.filter(author = request.user)
+        if queryset.count() >= 3:
+            return Response(status=status.HTTP_406_NOT_ACCEPTABLE)
         else:
-            return Response(status=status.HTTP_400_BAD_REQUEST)
+            serializer = ChatRoomSerializer(data=request.data)
+            if serializer.is_valid():
+                serializer.save(author = request.user)
+                return Response(status=status.HTTP_201_CREATED)
+            else:
+                return Response(status=status.HTTP_400_BAD_REQUEST)
     
     def delete(self, request, room_id):
-        room = get_object_or_404(ChatRoom, room_id=room_id)
+        room = get_object_or_404(ChatRoom, id=room_id)
+        check_participants = RoomChatParticipant.objects.filter(room_id=room_id)
+        # print(check_participants)
         if request.user == room.author:
-            room.delete()
-            return Response(status=status.HTTP_204_NO_CONTENT)
+            if not check_participants:
+                room.delete()
+                return Response(status=status.HTTP_204_NO_CONTENT)
+            else:
+                return Response(status=status.HTTP_403_FORBIDDEN)
         else:
             return Response(status=status.HTTP_401_UNAUTHORIZED)
